requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-environ==0.11.0
psycopg2-binary==2.9.6
drf-spectacular==0.28.1
python-decouple=3.8
